Add optional webhook server to runner to do continuous deployment.
If the runner is started with -p <port>, it will start a webhook server
on that port and watch for post-recieve POST requests formatted in
github fasion: http://help.github.com/post-receive-hooks/. By default,
it will watch the master branch though that can be changed by passing
the -b <branch> flag.

While this is useful during development, it's dangerous to go straight
to the runner for this -- the hook should call a CI server, which should
run tests, and when the tests pass, that should SIGHUP the runner.