Fix server lockup when uploading local badges. [closes #21]
I moved the user lookup code out of the middleware, into the app layer. While
in the middleware layer, it seemed to be causing a race condition where the
request events would fire off before the form handler could add listeners.
Moving the user lookup into the app layer is probably the right thing to do
anyway, but it does add another serving of callback spaghetti.