Remove all password related things from User model so we can drop bcrypt.
We aren't dealing with passwords or logins outside of browserid, so there
is no reason to keep the user password stuff. This also means there's no
reason to keep bcrypt as a dependency, which is good since removing it
means that all of our deps are now 100% js and uncompiled.