CSRF whitelisting authenticate route.
There's a problem in the issuing workflow that causes the CSRF token
check to fail. Since we are sending the data received by the
authentication route to browserid anyway for actual verification, there
is no security risk that I can see in disabling it until we figure out
the underlying issue.