Replace `emails` array in session instead of pushing.
This addresses issue #175. It was originally a `push` because we were
planning on having users be able to authenticate with multiple email
addresses to manage multiple identities within one backpack. This got
tabled, however, and we'll probably implement it in a slightly different
way. For now, let's leave the array but just replace it entirely every
time there's an authentication.