Allow strings & arrays for first param of `issue` API .
Addresses issue #167. When a string was being passed in before, it would
not immediately fail and instead see it as an array with `"string".length`
members which is dumb and wrong. Instead of making it error, I Just made
it more lenient so that the first argument can be a string or an array.