Initial integration with node-db-migrate.
This converts the CREATE TABLE code in lib/mysql.js to a migration
that can be rolled-back. Also added a test to ensure that it works.

Migrations are automatically run when the app is launched and when
the test database is set up at the beginning of each test. The latter
could get slow, but there are ways around it when that happens, such as
getting a mysqldump of the fully migrated state and executing it as a
series of SQL statements. But we'll cross that bridge if/when we come to it.