Added bin/db-migrate, which is a modified version of the db-migrate script.
I'm not a huge fan of this approach. Not only are we copy-pasting
code from db-migrate, thereby violating DRY, but the code uses undocumented
APIs which could break at any time. A better approach would be to
issue a pull request to node-db-migrate (or just use a fork) that
refactors it in a way that allows us to do what we need without
having to copy-paste the code. But figuring out how to do the refactoring
will take time, which we don't have much of right now.

Doing this also meant not calling createTables() in lib/mysql.js, i.e.
users need to run bin/db-migrate manually to ensure migration race
conditions don't occur between multiple server processes that are started
simultaneously.

I still need to:

* Document bin/db-migrate in the README, adding instructions on how to
  migrate up/down and create migrations.
* (optional) Add some code in app.js that checks to see whether we're
  at the latest migration, and report an error if not. This will be
  more helpful to developers than running code that expects a newer db
  schema and exploding confusingly.