Drop `type` column from badge table.
Right now, 100% of the badges are "hosted". Even when we get non-hosted
badges, the `type` can be inferred from whether the badge has an
`endpoint` or a `signature`, so it's more suited to being a method on
the model.

If we encounter the day when we need to do a superfast search on whether
a badge is signed or hosted, we should create an `index` table that has
that information foreign keyed against the relevant item in the badge
table.