New function library: analyzeAssertion
`analyzeAssertion` will be the wrapper for `openbadges-validator`. It
will deal with going out and grabbing the assertion if necessary and
normalizing any errors. It should be fairly lightweight and provide
detailed error responses so the consumers don't have to do much to
convert them into helpful user-facing errors.

Currently I've only tested 0.5.0 style assertions but theoretically it
will work with both as-is. I plan on implementing this in the codebase
instead of the various chained `remote` calls before I deal with testing
assertion version 1.0.0 support