Implement Badge#getFromBody
This is a safe method for getting deep items from the body of a badge
assertion. Instead of doing `badge.get('body').badge.issuer.origin` and
worrying that it's going to blow up if `issuer` is not a property of
`badge`, we can use `badge.getFromBody('badge.issuer.origin')` to get at
the deep propertites without having to do a bunch of &&'ing.