Added tests for publishBadge() to facebook lib.
However, this required changing the behavior of the function, because
its current behavior seems odd: it's calling an internal function,
publishGraphObject(), which parses the JSON response of facebook and
returns the id of the new open graph object. But then the caller
attempts to parse that id as JSON, and return it, which seems odd. So
I changed the function to just pass through the id. Will need to ask
@davelester if this is ok or not.