Fix fakeissuer.
I messed up in the assertions and had the verify type as signed for some
hosted assertions. The validator wasn't doing a great job of reporting
or stopping on that error, so it would fall through and hit the
`verify-hosted` error. I increased the error handling on
`openbadges-validator` so when this happens in the wild we get better
errors, and I fixed the problem here so we can actually test the
hashed/unhashed hosted badges.