Potential workaround for #932: force modal-less UI on IE.
This isn't a great workaround, as it boots the user off the issuer's site
and redirects them to the backpack without an easy way to return to where
they came from, but it's much better than what currently happens. And it
avoids the legal issues involved in delivering a P3P header that makes
the modal issuer flow "just work".